api/requirements.txt
fastapi
uvicorn
pandas
sklearn
azure-storage-blob==2.1.0
joblib